moved diff_processor helper functions to helper.py
